Struct aws_sdk_iotwireless::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for iotwireless.

This client allows ergonomic access to a iotwireless-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn associate_aws_account_with_partner_account(
&self
) -> AssociateAwsAccountWithPartnerAccount<C, M, R>

pub fn associate_wireless_device_with_thing(
&self
) -> AssociateWirelessDeviceWithThing<C, M, R>

pub fn associate_wireless_gateway_with_certificate(
&self
) -> AssociateWirelessGatewayWithCertificate<C, M, R>

pub fn associate_wireless_gateway_with_thing(
&self
) -> AssociateWirelessGatewayWithThing<C, M, R>

pub fn create_destination(&self) -> CreateDestination<C, M, R>

pub fn create_device_profile(&self) -> CreateDeviceProfile<C, M, R>

pub fn create_service_profile(&self) -> CreateServiceProfile<C, M, R>

pub fn create_wireless_device(&self) -> CreateWirelessDevice<C, M, R>

pub fn create_wireless_gateway(&self) -> CreateWirelessGateway<C, M, R>

pub fn create_wireless_gateway_task(&self) -> CreateWirelessGatewayTask<C, M, R>

pub fn create_wireless_gateway_task_definition(
&self
) -> CreateWirelessGatewayTaskDefinition<C, M, R>

pub fn delete_destination(&self) -> DeleteDestination<C, M, R>

pub fn delete_device_profile(&self) -> DeleteDeviceProfile<C, M, R>

pub fn delete_service_profile(&self) -> DeleteServiceProfile<C, M, R>

pub fn delete_wireless_device(&self) -> DeleteWirelessDevice<C, M, R>

pub fn delete_wireless_gateway(&self) -> DeleteWirelessGateway<C, M, R>

pub fn delete_wireless_gateway_task(&self) -> DeleteWirelessGatewayTask<C, M, R>

pub fn delete_wireless_gateway_task_definition(
&self
) -> DeleteWirelessGatewayTaskDefinition<C, M, R>

pub fn disassociate_aws_account_from_partner_account(
&self
) -> DisassociateAwsAccountFromPartnerAccount<C, M, R>

pub fn disassociate_wireless_device_from_thing(
&self
) -> DisassociateWirelessDeviceFromThing<C, M, R>

pub fn disassociate_wireless_gateway_from_certificate(
&self
) -> DisassociateWirelessGatewayFromCertificate<C, M, R>

pub fn disassociate_wireless_gateway_from_thing(
&self
) -> DisassociateWirelessGatewayFromThing<C, M, R>

pub fn get_destination(&self) -> GetDestination<C, M, R>

pub fn get_device_profile(&self) -> GetDeviceProfile<C, M, R>

pub fn get_log_levels_by_resource_types(
&self
) -> GetLogLevelsByResourceTypes<C, M, R>

pub fn get_partner_account(&self) -> GetPartnerAccount<C, M, R>

pub fn get_resource_log_level(&self) -> GetResourceLogLevel<C, M, R>

pub fn get_service_endpoint(&self) -> GetServiceEndpoint<C, M, R>

pub fn get_service_profile(&self) -> GetServiceProfile<C, M, R>

pub fn get_wireless_device(&self) -> GetWirelessDevice<C, M, R>

pub fn get_wireless_device_statistics(
&self
) -> GetWirelessDeviceStatistics<C, M, R>

pub fn get_wireless_gateway(&self) -> GetWirelessGateway<C, M, R>

pub fn get_wireless_gateway_certificate(
&self
) -> GetWirelessGatewayCertificate<C, M, R>

pub fn get_wireless_gateway_firmware_information(
&self
) -> GetWirelessGatewayFirmwareInformation<C, M, R>

pub fn get_wireless_gateway_statistics(
&self
) -> GetWirelessGatewayStatistics<C, M, R>

pub fn get_wireless_gateway_task(&self) -> GetWirelessGatewayTask<C, M, R>

pub fn get_wireless_gateway_task_definition(
&self
) -> GetWirelessGatewayTaskDefinition<C, M, R>

pub fn list_destinations(&self) -> ListDestinations<C, M, R>

pub fn list_device_profiles(&self) -> ListDeviceProfiles<C, M, R>

pub fn list_partner_accounts(&self) -> ListPartnerAccounts<C, M, R>

pub fn list_service_profiles(&self) -> ListServiceProfiles<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn list_wireless_devices(&self) -> ListWirelessDevices<C, M, R>

pub fn list_wireless_gateways(&self) -> ListWirelessGateways<C, M, R>

pub fn list_wireless_gateway_task_definitions(
&self
) -> ListWirelessGatewayTaskDefinitions<C, M, R>

pub fn put_resource_log_level(&self) -> PutResourceLogLevel<C, M, R>

pub fn reset_all_resource_log_levels(
&self
) -> ResetAllResourceLogLevels<C, M, R>

pub fn reset_resource_log_level(&self) -> ResetResourceLogLevel<C, M, R>

pub fn send_data_to_wireless_device(&self) -> SendDataToWirelessDevice<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn test_wireless_device(&self) -> TestWirelessDevice<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_destination(&self) -> UpdateDestination<C, M, R>

pub fn update_log_levels_by_resource_types(
&self
) -> UpdateLogLevelsByResourceTypes<C, M, R>

pub fn update_partner_account(&self) -> UpdatePartnerAccount<C, M, R>

pub fn update_wireless_device(&self) -> UpdateWirelessDevice<C, M, R>

pub fn update_wireless_gateway(&self) -> UpdateWirelessGateway<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self