Module aws_sdk_iotwireless::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

throttling_exception

See ThrottlingException

too_many_tags_exception

See TooManyTagsException

validation_exception

See ValidationException

Structs

AccessDeniedException

User does not have permission to perform this action.

AssociateAwsAccountWithPartnerAccountError

AssociateWirelessDeviceWithThingError

AssociateWirelessGatewayWithCertificateError

AssociateWirelessGatewayWithThingError

ConflictException

Adding, updating, or deleting the resource can cause an inconsistent state.

CreateDestinationError

CreateDeviceProfileError

CreateServiceProfileError

CreateWirelessDeviceError

CreateWirelessGatewayError

CreateWirelessGatewayTaskDefinitionError

CreateWirelessGatewayTaskError

DeleteDestinationError

DeleteDeviceProfileError

DeleteServiceProfileError

DeleteWirelessDeviceError

DeleteWirelessGatewayError

DeleteWirelessGatewayTaskDefinitionError

DeleteWirelessGatewayTaskError

DisassociateAwsAccountFromPartnerAccountError

DisassociateWirelessDeviceFromThingError

DisassociateWirelessGatewayFromCertificateError

DisassociateWirelessGatewayFromThingError

GetDestinationError

GetDeviceProfileError

GetLogLevelsByResourceTypesError

GetPartnerAccountError

GetResourceLogLevelError

GetServiceEndpointError

GetServiceProfileError

GetWirelessDeviceError

GetWirelessDeviceStatisticsError

GetWirelessGatewayCertificateError

GetWirelessGatewayError

GetWirelessGatewayFirmwareInformationError

GetWirelessGatewayStatisticsError

GetWirelessGatewayTaskDefinitionError

GetWirelessGatewayTaskError

InternalServerException

An unexpected error occurred while processing a request.

ListDestinationsError

ListDeviceProfilesError

ListPartnerAccountsError

ListServiceProfilesError

ListTagsForResourceError

ListWirelessDevicesError

ListWirelessGatewayTaskDefinitionsError

ListWirelessGatewaysError

PutResourceLogLevelError

ResetAllResourceLogLevelsError

ResetResourceLogLevelError

ResourceNotFoundException

Resource does not exist.

SendDataToWirelessDeviceError

TagResourceError

TestWirelessDeviceError

ThrottlingException

The request was denied because it exceeded the allowed API request rate.

TooManyTagsException

The request was denied because the resource can't have any more tags.

UntagResourceError

UpdateDestinationError

UpdateLogLevelsByResourceTypesError

UpdatePartnerAccountError

UpdateWirelessDeviceError

UpdateWirelessGatewayError

ValidationException

The input did not meet the specified constraints.

Enums

AssociateAwsAccountWithPartnerAccountErrorKind

AssociateWirelessDeviceWithThingErrorKind

AssociateWirelessGatewayWithCertificateErrorKind

AssociateWirelessGatewayWithThingErrorKind

CreateDestinationErrorKind

CreateDeviceProfileErrorKind

CreateServiceProfileErrorKind

CreateWirelessDeviceErrorKind

CreateWirelessGatewayErrorKind

CreateWirelessGatewayTaskDefinitionErrorKind

CreateWirelessGatewayTaskErrorKind

DeleteDestinationErrorKind

DeleteDeviceProfileErrorKind

DeleteServiceProfileErrorKind

DeleteWirelessDeviceErrorKind

DeleteWirelessGatewayErrorKind

DeleteWirelessGatewayTaskDefinitionErrorKind

DeleteWirelessGatewayTaskErrorKind

DisassociateAwsAccountFromPartnerAccountErrorKind

DisassociateWirelessDeviceFromThingErrorKind

DisassociateWirelessGatewayFromCertificateErrorKind

DisassociateWirelessGatewayFromThingErrorKind

GetDestinationErrorKind

GetDeviceProfileErrorKind

GetLogLevelsByResourceTypesErrorKind

GetPartnerAccountErrorKind

GetResourceLogLevelErrorKind

GetServiceEndpointErrorKind

GetServiceProfileErrorKind

GetWirelessDeviceErrorKind

GetWirelessDeviceStatisticsErrorKind

GetWirelessGatewayCertificateErrorKind

GetWirelessGatewayErrorKind

GetWirelessGatewayFirmwareInformationErrorKind

GetWirelessGatewayStatisticsErrorKind

GetWirelessGatewayTaskDefinitionErrorKind

GetWirelessGatewayTaskErrorKind

ListDestinationsErrorKind

ListDeviceProfilesErrorKind

ListPartnerAccountsErrorKind

ListServiceProfilesErrorKind

ListTagsForResourceErrorKind

ListWirelessDevicesErrorKind

ListWirelessGatewayTaskDefinitionsErrorKind

ListWirelessGatewaysErrorKind

PutResourceLogLevelErrorKind

ResetAllResourceLogLevelsErrorKind

ResetResourceLogLevelErrorKind

SendDataToWirelessDeviceErrorKind

TagResourceErrorKind

TestWirelessDeviceErrorKind

UntagResourceErrorKind

UpdateDestinationErrorKind

UpdateLogLevelsByResourceTypesErrorKind

UpdatePartnerAccountErrorKind

UpdateWirelessDeviceErrorKind

UpdateWirelessGatewayErrorKind