Module aws_sdk_iotwireless::model

Modules

abp_v10_x

See AbpV10X

abp_v11

See AbpV11

certificate_list

See CertificateList

destinations

See Destinations

device_profile

See DeviceProfile

lo_ra_wan_device

See LoRaWanDevice

lo_ra_wan_device_metadata

See LoRaWanDeviceMetadata

lo_ra_wan_device_profile

See LoRaWanDeviceProfile

lo_ra_wan_gateway

See LoRaWanGateway

lo_ra_wan_gateway_current_version

See LoRaWanGatewayCurrentVersion

lo_ra_wan_gateway_metadata

See LoRaWanGatewayMetadata

lo_ra_wan_gateway_version

See LoRaWanGatewayVersion

lo_ra_wan_get_service_profile_info

See LoRaWanGetServiceProfileInfo

lo_ra_wan_list_device

See LoRaWanListDevice

lo_ra_wan_send_data_to_device

See LoRaWanSendDataToDevice

lo_ra_wan_service_profile

See LoRaWanServiceProfile

lo_ra_wan_update_device

See LoRaWanUpdateDevice

lo_ra_wan_update_gateway_task_create

See LoRaWanUpdateGatewayTaskCreate

lo_ra_wan_update_gateway_task_entry

See LoRaWanUpdateGatewayTaskEntry

otaa_v10_x

See OtaaV10X

otaa_v11

See OtaaV11

service_profile

See ServiceProfile

session_keys_abp_v10_x

See SessionKeysAbpV10X

session_keys_abp_v11

See SessionKeysAbpV11

sidewalk_account_info

See SidewalkAccountInfo

sidewalk_account_info_with_fingerprint

See SidewalkAccountInfoWithFingerprint

sidewalk_device

See SidewalkDevice

sidewalk_device_metadata

See SidewalkDeviceMetadata

sidewalk_list_device

See SidewalkListDevice

sidewalk_send_data_to_device

See SidewalkSendDataToDevice

sidewalk_update_account

See SidewalkUpdateAccount

tag

See Tag

update_wireless_gateway_task_create

See UpdateWirelessGatewayTaskCreate

update_wireless_gateway_task_entry

See UpdateWirelessGatewayTaskEntry

wireless_device_event_log_option

See WirelessDeviceEventLogOption

wireless_device_log_option

See WirelessDeviceLogOption

wireless_device_statistics

See WirelessDeviceStatistics

wireless_gateway_event_log_option

See WirelessGatewayEventLogOption

wireless_gateway_log_option

See WirelessGatewayLogOption

wireless_gateway_statistics

See WirelessGatewayStatistics

wireless_metadata

See WirelessMetadata

Structs

AbpV10X

ABP device object for LoRaWAN specification v1.0.x

AbpV11

ABP device object for LoRaWAN specification v1.1

CertificateList

List of sidewalk certificates.

Destinations

Describes a destination.

DeviceProfile

Describes a device profile.

LoRaWanDevice

LoRaWAN object for create functions.

LoRaWanDeviceMetadata

LoRaWAN device metatdata.

LoRaWanDeviceProfile

LoRaWANDeviceProfile object.

LoRaWanGateway

LoRaWANGateway object.

LoRaWanGatewayCurrentVersion

LoRaWANGatewayCurrentVersion object.

LoRaWanGatewayMetadata

LoRaWAN gateway metatdata.

LoRaWanGatewayVersion

LoRaWANGatewayVersion object.

LoRaWanGetServiceProfileInfo

LoRaWANGetServiceProfileInfo object.

LoRaWanListDevice

LoRaWAN object for list functions.

LoRaWanSendDataToDevice

LoRaWAN router info.

LoRaWanServiceProfile

LoRaWANServiceProfile object.

LoRaWanUpdateDevice

LoRaWAN object for update functions.

LoRaWanUpdateGatewayTaskCreate

LoRaWANUpdateGatewayTaskCreate object.

LoRaWanUpdateGatewayTaskEntry

LoRaWANUpdateGatewayTaskEntry object.

OtaaV10X

OTAA device object for v1.0.x

OtaaV11

OTAA device object for v1.1

ServiceProfile

Information about a service profile.

SessionKeysAbpV10X

Session keys for ABP v1.1

SessionKeysAbpV11

Session keys for ABP v1.1

SidewalkAccountInfo

Information about a Sidewalk account.

SidewalkAccountInfoWithFingerprint

Information about a Sidewalk account.

SidewalkDevice

Sidewalk device object.

SidewalkDeviceMetadata

MetaData for Sidewalk device.

SidewalkListDevice

Sidewalk object used by list functions.

SidewalkSendDataToDevice

Information about a Sidewalk router.

SidewalkUpdateAccount

Sidewalk update.

Tag

A simple label consisting of a customer-defined key-value pair

UpdateWirelessGatewayTaskCreate

UpdateWirelessGatewayTaskCreate object.

UpdateWirelessGatewayTaskEntry

UpdateWirelessGatewayTaskEntry object.

WirelessDeviceEventLogOption

The log options for a wireless device event and can be used to set log levels for a specific wireless device event.

WirelessDeviceLogOption

The log options for wireless devices and can be used to set log levels for a specific type of wireless device.

WirelessDeviceStatistics

Information about a wireless device's operation.

WirelessGatewayEventLogOption

The log options for a wireless gateway event and can be used to set log levels for a specific wireless gateway event.

WirelessGatewayLogOption

The log options for wireless gateways and can be used to set log levels for a specific type of wireless gateway.

WirelessGatewayStatistics

Information about a wireless gateway's operation.

WirelessMetadata

WirelessMetadata object.

Enums

BatteryLevel

Sidewalk device battery level.

ConnectionStatus

DeviceState

Device state defines the device status of sidewalk device.

Event

Sidewalk device status notification.

ExpressionType

LogLevel

The log level for a log message.

MessageType

Sidewalk device message type. Default value is CUSTOM_COMMAND_ID_NOTIFY.

PartnerType

SigningAlg

The certificate chain algorithm provided by sidewalk.

WirelessDeviceEvent

The event for a log message, if the log message is tied to a wireless device.

WirelessDeviceIdType

WirelessDeviceType

WirelessGatewayEvent

The event for a log message, if the log message is tied to a wireless gateway.

WirelessGatewayIdType

WirelessGatewayServiceType

WirelessGatewayTaskDefinitionType

WirelessGatewayTaskStatus

WirelessGatewayType

The wireless gateway type.