Struct aws_sdk_iotwireless::model::lo_ra_wan_get_service_profile_info::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for LoRaWanGetServiceProfileInfo

Implementations

impl Builder

pub fn ul_rate(self, input: i32) -> Self

The ULRate value.

pub fn set_ul_rate(self, input: Option<i32>) -> Self

pub fn ul_bucket_size(self, input: i32) -> Self

The ULBucketSize value.

pub fn set_ul_bucket_size(self, input: Option<i32>) -> Self

pub fn ul_rate_policy(self, input: impl Into<String>) -> Self

The ULRatePolicy value.

pub fn set_ul_rate_policy(self, input: Option<String>) -> Self

pub fn dl_rate(self, input: i32) -> Self

The DLRate value.

pub fn set_dl_rate(self, input: Option<i32>) -> Self

pub fn dl_bucket_size(self, input: i32) -> Self

The DLBucketSize value.

pub fn set_dl_bucket_size(self, input: Option<i32>) -> Self

pub fn dl_rate_policy(self, input: impl Into<String>) -> Self

The DLRatePolicy value.

pub fn set_dl_rate_policy(self, input: Option<String>) -> Self

pub fn add_gw_metadata(self, input: bool) -> Self

The AddGWMetaData value.

pub fn set_add_gw_metadata(self, input: Option<bool>) -> Self

pub fn dev_status_req_freq(self, input: i32) -> Self

The DevStatusReqFreq value.

pub fn set_dev_status_req_freq(self, input: Option<i32>) -> Self

pub fn report_dev_status_battery(self, input: bool) -> Self

The ReportDevStatusBattery value.

pub fn set_report_dev_status_battery(self, input: Option<bool>) -> Self

pub fn report_dev_status_margin(self, input: bool) -> Self

The ReportDevStatusMargin value.

pub fn set_report_dev_status_margin(self, input: Option<bool>) -> Self

pub fn dr_min(self, input: i32) -> Self

The DRMin value.

pub fn set_dr_min(self, input: Option<i32>) -> Self

pub fn dr_max(self, input: i32) -> Self

The DRMax value.

pub fn set_dr_max(self, input: Option<i32>) -> Self

pub fn channel_mask(self, input: impl Into<String>) -> Self

The ChannelMask value.

pub fn set_channel_mask(self, input: Option<String>) -> Self

pub fn pr_allowed(self, input: bool) -> Self

The PRAllowed value that describes whether passive roaming is allowed.

pub fn set_pr_allowed(self, input: Option<bool>) -> Self

pub fn hr_allowed(self, input: bool) -> Self

The HRAllowed value that describes whether handover roaming is allowed.

pub fn set_hr_allowed(self, input: Option<bool>) -> Self

pub fn ra_allowed(self, input: bool) -> Self

The RAAllowed value that describes whether roaming activation is allowed.

pub fn set_ra_allowed(self, input: Option<bool>) -> Self

pub fn nwk_geo_loc(self, input: bool) -> Self

The NwkGeoLoc value.

pub fn set_nwk_geo_loc(self, input: Option<bool>) -> Self

pub fn target_per(self, input: i32) -> Self

The TargetPER value.

pub fn set_target_per(self, input: Option<i32>) -> Self

pub fn min_gw_diversity(self, input: i32) -> Self

The MinGwDiversity value.

pub fn set_min_gw_diversity(self, input: Option<i32>) -> Self

pub fn build(self) -> LoRaWanGetServiceProfileInfo

Consumes the builder and constructs a LoRaWanGetServiceProfileInfo