Module aws_sdk_iotwireless::operation

Structs

AssociateAwsAccountWithPartnerAccount

Associates a partner account with your AWS account.

AssociateWirelessDeviceWithThing

Associates a wireless device with a thing.

AssociateWirelessGatewayWithCertificate

Associates a wireless gateway with a certificate.

AssociateWirelessGatewayWithThing

Associates a wireless gateway with a thing.

CreateDestination

Creates a new destination that maps a device message to an AWS IoT rule.

CreateDeviceProfile

Creates a new device profile.

CreateServiceProfile

Creates a new service profile.

CreateWirelessDevice

Provisions a wireless device.

CreateWirelessGateway

Provisions a wireless gateway.

CreateWirelessGatewayTask

Creates a task for a wireless gateway.

CreateWirelessGatewayTaskDefinition

Creates a gateway task definition.

DeleteDestination

Deletes a destination.

DeleteDeviceProfile

Deletes a device profile.

DeleteServiceProfile

Deletes a service profile.

DeleteWirelessDevice

Deletes a wireless device.

DeleteWirelessGateway

Deletes a wireless gateway.

DeleteWirelessGatewayTask

Deletes a wireless gateway task.

DeleteWirelessGatewayTaskDefinition

Deletes a wireless gateway task definition. Deleting this task definition does not affect tasks that are currently in progress.

DisassociateAwsAccountFromPartnerAccount

Disassociates your AWS account from a partner account. If PartnerAccountId and PartnerType are null, disassociates your AWS account from all partner accounts.

DisassociateWirelessDeviceFromThing

Disassociates a wireless device from its currently associated thing.

DisassociateWirelessGatewayFromCertificate

Disassociates a wireless gateway from its currently associated certificate.

DisassociateWirelessGatewayFromThing

Disassociates a wireless gateway from its currently associated thing.

GetDestination

Gets information about a destination.

GetDeviceProfile

Gets information about a device profile.

GetLogLevelsByResourceTypes

Returns current default log levels or log levels by resource types. Based on resource types, log levels can be

GetPartnerAccount

Gets information about a partner account. If PartnerAccountId and PartnerType are null, returns all partner accounts.

GetResourceLogLevel

Fetches the log-level override, if any, for a given resource-ID and resource-type. It can be used for

GetServiceEndpoint

Gets the account-specific endpoint for Configuration and Update Server (CUPS) protocol or LoRaWAN Network Server (LNS) connections.

GetServiceProfile

Gets information about a service profile.

GetWirelessDevice

Gets information about a wireless device.

GetWirelessDeviceStatistics

Gets operating information about a wireless device.

GetWirelessGateway

Gets information about a wireless gateway.

GetWirelessGatewayCertificate

Gets the ID of the certificate that is currently associated with a wireless gateway.

GetWirelessGatewayFirmwareInformation

Gets the firmware version and other information about a wireless gateway.

GetWirelessGatewayStatistics

Gets operating information about a wireless gateway.

GetWirelessGatewayTask

Gets information about a wireless gateway task.

GetWirelessGatewayTaskDefinition

Gets information about a wireless gateway task definition.

ListDestinations

Lists the destinations registered to your AWS account.

ListDeviceProfiles

Lists the device profiles registered to your AWS account.

ListPartnerAccounts

Lists the partner accounts associated with your AWS account.

ListServiceProfiles

Lists the service profiles registered to your AWS account.

ListTagsForResource

Lists the tags (metadata) you have assigned to the resource.

ListWirelessDevices

Lists the wireless devices registered to your AWS account.

ListWirelessGatewayTaskDefinitions

List the wireless gateway tasks definitions registered to your AWS account.

ListWirelessGateways

Lists the wireless gateways registered to your AWS account.

PutResourceLogLevel

Sets the log-level override for a resource-ID and resource-type. This option can be specified for a wireless gateway

ResetAllResourceLogLevels

Removes the log-level overrides for all resources; both wireless devices and wireless gateways.

ResetResourceLogLevel

Removes the log-level override, if any, for a specific resource-ID and resource-type. It can be used for

SendDataToWirelessDevice

Sends a decrypted application data frame to a device.

TagResource

Adds a tag to a resource.

TestWirelessDevice

Simulates a provisioned device by sending an uplink data payload of Hello.

UntagResource

Removes one or more tags from a resource.

UpdateDestination

Updates properties of a destination.

UpdateLogLevelsByResourceTypes

Set default log level, or log levels by resource types. This can be for wireless device log options or

UpdatePartnerAccount

Updates properties of a partner account.

UpdateWirelessDevice

Updates properties of a wireless device.

UpdateWirelessGateway

Updates properties of a wireless gateway.